Refactored validators to make it easier to extend with additional checks. Added function re-definition fix and tests.
