+++ client/testFixture/Test.4dm
@@ -28,7 +28,12 @@ void My_function(Text_Text_Map map)
 };
 // -----------------------------------------------------------
 //
-
+Integer create_rgb(Integer r,Integer g,Integer b)
+// --------------------------------------------------------------------------
+// --------------------------------------------------------------------------
+{
+  return((1 << 31) | (r << 16) | (g << 8) | b);
+}
 //                  4  SC
 //                  5  CS
 //                  6  HIP
